training/requirements.txt
numpy==1.24.0
tensorflow==2.13.0
scikit-learn==1.0.2
joblib==1.4.2
pillow==11.1.0
